requirements.txt
# Core orchestration
langgraph>=0.2.28
langchain>=0.2.16
langchain-core>=0.2.38
langchain-anthropic>=0.1.23
# langchain-openai not needed: Anthropic is the primary LLM on HF Spaces

# Data & schemas
pydantic>=2.7.0
python-dotenv>=1.0.0

# Memory (chromadb is optional — VectorStore gracefully degrades without it)
# chromadb>=0.5.5  # loads ~300MB ONNX model, causes OOM on HF Spaces free tier

# CLI
typer>=0.12.5
rich>=13.7.1

# Rendering
pillow>=10.4.0

# fal.ai T2V
fal-client>=0.5.0

# Music generation
replicate>=0.34.0

# Utilities
httpx>=0.27.0

# Web server
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Auth & Billing
python-jose[cryptography]>=3.3.0
stripe>=7.0.0
